Dev_Synergy/accident_detectionv1/requirements.txt
pandas
numpy
joblib
requests
smbus2
pyserial
tflite-runtime
scikit-learn
